refactor: add new directory for store and move file
